+++ ConsoleApp1/Interfaces/ICalc.cs
@@ -4,5 +4,5 @@ public interface ICalc
 {
     int Add(int a, int b);
     int Subtract(int a, int b);
-    double Divide(double a, double b);
+    int Divide(int a, int b);
 }
